Create R square value differences for weighted least squares
using the squared correlation between the observed values of the response and the WLS fitted values (reweighted for comparability with those observations),  , the fraction of the Total Sum of Squares of the observed values of the response, which is explained by the fit, and , the multiple R-square from the OLS analysis. finding R square in each of these cases.