Create two stage least squares model for serial correlation
A 2 stage least squares model after specifying model errors being dependent on laggesd variables and determining the least squares estimator of the AR (1) model. This is then plugged into the linear regression model, for which the parameters remain the same (except beta 0).
This can be used to detect and modify a regression model for serial correlation. In this case, the least squares parameter for the error dependence is estimated on the residuals rather than the errors of the model as these are unobservable.