Detecting influential observations graphically in IMF2015 data
Plotting influential measures graphically, using the plots of:
1) Cook's distance with its Index
2) DFFITs with its index
3) DFBETAs with its index
4) COVRATIO and its index
5) Standardized Residuals vs Leverages
6) Leverages vs Cook's Distance (Curves)
7) Leverages vs Cook's Distance (With Residual Curves)

Using the Lm object as an argument in plot(), we use the which function to plot these observations (there are 6 types of plots for regression diagnostics in total), where we mention and plot the relevant thresholds.
Excess objects are pruned. Greece seems to be an extreme influence variable.