Ridge regression without parameterized penalty variable
running ridge regression on the dataset of nhospital. Ridge regression on the standardized predictor variables with penalty parameter λ=5000, the norm of the estimated ridge regression coefficients (without intercept) and the norm of the estimated least square coefficients (without intercept).
Here, the value of lambda is prefit to 5000.

Ideally, variables should be standardized before being used in the design matrix of a ridge or a lasso regression.